requirements.txt
streamlit==1.32.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==5.12.4
pandas==2.2.1
python-dotenv==1.0.1 